+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 #2.2.9
 - Improved 'populate sidebar with regular results' functionality
+- Added option to configure populated results position
 - Improved favicons search
 - Added loading spinner when loading favicons
 - Favicons are now aligned with title
 - Fixed for 'Apply style to widgets' not disabling correctly
-- Small performance and layout improvements
+- Performance and layout improvements
 
 #2.2.8 
 - Improved short header calculation